Moved pure wasm stuff to another folder
