+++ src/wasm/fastx_wasm.rs
